TP terminado

+++ TP1/conversion.c
@@ -1,5 +1,5 @@
 #include"conversion.h"
 
-int convertir (int valorAconvertir){
+float convertir (float valorAconvertir){
     return valorAconvertir*1.852;
 }
