+++ TP1/conversion.h
@@ -1,4 +1,4 @@
 #ifndef CONVERSION_H_INCLUDED
 #define CONVERSION_H_INCLUDED
-int convertir (int valorAconvertir);
+float convertir (float valorAconvertir);
 #endif //CONVERSION_H_INCLUDED
